+++ activestorage/test/dummy/config/initializers/cookies.rb
@@ -5,3 +5,7 @@
 # Specify a serializer for the signed and encrypted cookie jars.
 # Valid options are :json, :marshal, and :hybrid.
 Rails.application.config.action_dispatch.cookies_serializer = :json
+
+# Specify the SameSite level protection for the cookies
+# Valid options are :no_protection, :lax, and :strict.
+Rails.application.config.action_dispatch.cookies_same_site_protection = :no_protection